fix: should be optional in initial
