+++ packages/plugins/theme-editor/src/client/hooks/useThemeSettings.tsx
@@ -1,4 +1,4 @@
-import { useCurrentUserContext, useSystemSettings } from '@nocobase/client';
+import { useAPIClient, useCurrentUserContext, useSystemSettings } from '@nocobase/client';
 import { error } from '@nocobase/utils/client';
 import { MenuProps, Select } from 'antd';
 import React, { useEffect, useMemo } from 'react';
@@ -21,9 +21,11 @@ function Label() {
   const { t } = useTranslation();
   const currentUser = useCurrentUserContext();
   const systemSettings = useSystemSettings();
-  const { run, error: err, data } = useThemeListContext();
-  const { updateUserThemeSettings } = useUpdateThemeSettings();
+  const { run, error: err, data, refresh } = useThemeListContext();
+  const { updateUserThemeSettings, updateSystemThemeSettings } = useUpdateThemeSettings();
   const currentThemeId = useCurrentThemeId();
+  const themeId = useCurrentThemeId();
+  const api = useAPIClient();
 
   const options = useMemo(() => {
     return data
@@ -34,13 +36,32 @@ function Label() {
           value: item.id,
         };
       });
-  }, [data]);
+  }, [data, t]);
 
   useEffect(() => {
     if (!data) {
       run();
     }
   }, []);
+
+  useEffect(() => {
+    const init = async () => {
+      // 当 themeId 为空时表示插件是第一次被启用
+      if (themeId == null && data) {
+        const firstTheme = data[0];
+
+        try {
+          // 避免并发请求，在本地存储中容易出问题
+          await updateSystemThemeSettings(firstTheme.id);
+          await updateUserThemeSettings(firstTheme.id);
+        } catch (err) {
+          error(err);
+        }
+      }
+    };
+
+    init();
+  }, [themeId, updateSystemThemeSettings, updateUserThemeSettings, data, api, refresh]);
 
   if (err) {
     error(err);
